+++ samples/instrumentation-quickstart/setup_opentelemetry.py
@@ -14,13 +14,11 @@
 
 import os
 
-from opentelemetry import _events as events
 from opentelemetry import _logs as logs
 from opentelemetry import metrics, trace
 from opentelemetry.exporter.otlp.proto.http._log_exporter import OTLPLogExporter
 from opentelemetry.exporter.otlp.proto.http.metric_exporter import OTLPMetricExporter
 from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
-from opentelemetry.sdk._events import EventLoggerProvider
 from opentelemetry.sdk._logs import LoggerProvider
 from opentelemetry.sdk._logs.export import BatchLogRecordProcessor
 from opentelemetry.sdk.metrics import MeterProvider
@@ -28,7 +26,6 @@ from opentelemetry.sdk.metrics.export import PeriodicExportingMetricReader
 from opentelemetry.sdk.resources import SERVICE_INSTANCE_ID, Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor
-
 
 # [START opentelemetry_instrumentation_setup_opentelemetry]
 def setup_opentelemetry() -> None:
@@ -39,7 +36,6 @@ def setup_opentelemetry() -> None:
             SERVICE_INSTANCE_ID: f"worker-{os.getpid()}",
         }
     )
-
     # Set up OpenTelemetry Python SDK
     tracer_provider = TracerProvider(resource=resource)
     tracer_provider.add_span_processor(BatchSpanProcessor(OTLPSpanExporter()))
@@ -48,9 +44,6 @@ def setup_opentelemetry() -> None:
     logger_provider = LoggerProvider(resource=resource)
     logger_provider.add_log_record_processor(BatchLogRecordProcessor(OTLPLogExporter()))
     logs.set_logger_provider(logger_provider)
-
-    event_logger_provider = EventLoggerProvider(logger_provider)
-    events.set_event_logger_provider(event_logger_provider)
 
     reader = PeriodicExportingMetricReader(OTLPMetricExporter())
     meter_provider = MeterProvider(metric_readers=[reader], resource=resource)